Fixed some GCC warnings: -Wconversion -Wshadow -Wunused-parameter. JIRA:CXX-5413
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@74562 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/seqfeat/Trna_ext.cpp
@@ -76,8 +76,8 @@ bool CTrna_ext::ParseDegenerateCodon(CTrna_ext & tRNA, const string & codon)
         return false;
     }
 
-    int idx = intToChr.find(codon[2]);
-    if (idx == (int)string::npos) return false;
+    size_t idx = intToChr.find(codon[2]);
+    if (idx == string::npos) return false;
 
     const char *expanded_codon_letter = codonLetterExpand[idx];
     const char *iter = expanded_codon_letter;
